moved graph logic from playerInfo to ratingGraph compo
